pyproject.toml
[tool.poetry]
name = "HealthSyncIQ"
version = "0.1.0"
description = "A RBAC Hospital Management Platform"
authors = ["DavidHODs <davidoluwatobi41@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0" 
fastapi = ">=0.110.0,<1.0.0"
uvicorn = { extras = ["standard"], version = ">=0.29.0,<1.0.0" }
pydantic = {extras = ["email"], version = "^2.11.4"}
sqlalchemy = ">=2.0.0,<3.0.0"
alembic = ">=1.13.1,<2.0.0"
httpx = ">=0.27.0,<1.0.0"
gunicorn = ">=22.0.0,<23.0.0"
psycopg = { extras = ["binary"], version = ">=3.1.0,<4.0.0" }
python-dotenv = "^1.0.1"
autopep8 = "^2.3.2"
isort = "^6.0.1"
mypy = "^1.15.0"
ruff = "^0.11.0"
psycopg2 = "^2.9.10"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=3.7.0,<4.0.0"

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
strict = true
explicit_package_bases = true

[tool.autopep8]
indent-size = 2
max-line-length = 80

[tool.isort]
profile = "black"

[tool.ruff]
fix = true
line-length = 80

[tool.ruff.lint]
select = ["F401", "F841"] 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
